# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Balances and Transactions API",
    "description": "This API provides endpoints to get the current balance and a list of transactions.",
    "version": "1.0.0",
    "servers": {
      "url": "https://uh4goxppjc7stkg24d6fdma4t40wxtly.lambda-url.eu-central-1.on.aws"
    }
  },
  "paths": {
    "/balances": {
      "get": {
        "security": [
          {
            "xApiKey": []
          }
        ],
        "description": "Get the current balance.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "amount": {
                      "type": "number"
                    },
                    "currency": {
                      "type": "string",
                      "enum": [
                        "EUR"
                      ]
                    },
                    "date": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transactions": {
      "get": {
        "security": [
          {
            "xApiKey": []
          }
        ],
        "description": "Get a list of transactions.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transactions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "amount": {
                            "type": "number"
                          },
                          "currency": {
                            "type": "string",
                            "enum": [
                              "EUR"
                            ]
                          },
                          "date": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "BOOKED",
                              "PROCESSED",
                              "CANCELLED"
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/historical-balances": {
      "get": {
        "security": [
          {
            "xApiKey": []
          }
        ],
        "description": "Get a list of daily balances for a specific date range.",
        "parameters": [{
          "name": "from",
          "in": "query",
          "description": "Start date of the date range",
          "required": "true",
          "type": "string",
          "example": "2022-05-04"
        },
        {
          "name": "to",
          "in": "query",
          "description": "End date of the date range",
          "required": "true",
          "type": "string",
          "example": "2022-05-06"
        },
        {
          "name": "sort",
          "in": "query",
          "description": "Sorting order of daily balances: desc or asc",
          "type": "string",
          "example": "desc"
        }
      ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transactions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": {
                            "type": "string",
                            "format": "date"
                          },
                          "amount": {
                            "type": "number"
                          },
                          "currency": {
                            "type": "string",
                            "enum": [
                              "EUR"
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Invalid request, this route doesn't exist"
          },
          "400": {
            "description": "Invalid/missing dates parameters or an invalid sorting order"
          },
          "502": {
            "description": "Couldn't get data from Banxware API"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "xApiKey": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header"
      }
    }
  }
}